+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog and versioning
 
 - Add MQLContext to capture addition query information that cannot be expressed in MQL string
     - Add visitors to support MQLContext serialization
+- Support curried functions in the MQL grammar
 
 
 2.0.10